Warnings: reduce warnigs (level 4) to ~40  * Removed unused console option.  * map/area flags are not  * removed negative/unused value in all function calls for RemoveAllAurasByMechanic.

+++ src/collision/BoundingIntervalHierarchy.cpp
@@ -113,7 +113,7 @@ void BIH::subdivide(int left, int right, std::vector<uint32> &tempTree, buildDat
             if (1.3f * nodeNewW < nodeBoxW)
             {
                 stats.updateBVH2();
-                int nextIndex = tempTree.size();
+                int nextIndex = static_cast<int>(tempTree.size());
                 // allocate child
                 tempTree.push_back(0);
                 tempTree.push_back(0);
@@ -177,7 +177,7 @@ void BIH::subdivide(int left, int right, std::vector<uint32> &tempTree, buildDat
             {
                 // second time through - lets create the previous split
                 // since it produced empty space
-                int nextIndex = tempTree.size();
+                int nextIndex = static_cast<int>(tempTree.size());
                 // allocate child node
                 tempTree.push_back(0);
                 tempTree.push_back(0);
@@ -207,7 +207,7 @@ void BIH::subdivide(int left, int right, std::vector<uint32> &tempTree, buildDat
         }
     }
     // compute index of child nodes
-    int nextIndex = tempTree.size();
+    int nextIndex = static_cast<int>(tempTree.size());
     // allocate left node
     int nl = right - left + 1;
     int nr = rightOrig - (right + 1) + 1;
@@ -247,13 +247,13 @@ void BIH::subdivide(int left, int right, std::vector<uint32> &tempTree, buildDat
 
 bool BIH::writeToFile(FILE* wf) const
 {
-    uint32 treeSize = tree.size();
+    uint32 treeSize = static_cast<uint32>(tree.size());
     size_t check = 0;
     check += fwrite(&bounds.low(), sizeof(float), 3, wf);
     check += fwrite(&bounds.high(), sizeof(float), 3, wf);
     check += fwrite(&treeSize, sizeof(uint32), 1, wf);
     check += fwrite(&tree[0], sizeof(uint32), treeSize, wf);
-    uint32 count = objects.size();
+    uint32 count = static_cast<uint32>(objects.size());
     check += fwrite(&count, sizeof(uint32), 1, wf);
     check += fwrite(&objects[0], sizeof(uint32), count, wf);
     return (check == (3 + 3 + 2 + treeSize + count));
@@ -264,15 +264,15 @@ bool BIH::readFromFile(FILE* rf)
     uint32 treeSize;
     G3D::Vector3 lo, hi;
     uint32 check = 0, count = 0;
-    check += fread(&lo, sizeof(float), 3, rf);
-    check += fread(&hi, sizeof(float), 3, rf);
+    check += static_cast<uint32>(fread(&lo, sizeof(float), 3, rf));
+    check += static_cast<uint32>(fread(&hi, sizeof(float), 3, rf));
     bounds = G3D::AABox(lo, hi);
-    check += fread(&treeSize, sizeof(uint32), 1, rf);
+    check += static_cast<uint32>(fread(&treeSize, sizeof(uint32), 1, rf));
     tree.resize(treeSize);
-    check += fread(&tree[0], sizeof(uint32), treeSize, rf);
-    check += fread(&count, sizeof(uint32), 1, rf);
+    check += static_cast<uint32>(fread(&tree[0], sizeof(uint32), treeSize, rf));
+    check += static_cast<uint32>(fread(&count, sizeof(uint32), 1, rf));
     objects.resize(count); // = new uint32[nObjects];
-    check += fread(&objects[0], sizeof(uint32), count, rf);
+    check += static_cast<uint32>(fread(&objects[0], sizeof(uint32), count, rf));
     return (uint64(check) == uint64(3 + 3 + 1 + 1 + uint64(treeSize) + uint64(count)));
 }
